+++ src/ruleLoader.ts
@@ -18,13 +18,15 @@
 import * as fs from "fs";
 import * as path from "path";
 
-import {getRulesDirectories} from "./configuration";
-import {IDisabledInterval, IRule} from "./language/rule/rule";
-import {camelize, dedent} from "./utils";
+import { getRelativePath } from "./configuration";
+import { AbstractRule } from "./language/rule/abstractRule";
+import { IDisabledInterval, IRule } from "./language/rule/rule";
+import { arrayify, camelize, dedent } from "./utils";
 
 const moduleDirectory = path.dirname(module.filename);
 const CORE_RULES_DIRECTORY = path.resolve(moduleDirectory, ".", "rules");
 const shownDeprecations: string[] = [];
+const cachedRules = new Map<string, typeof AbstractRule | null>(); // null indicates that the rule was not found
 
 export interface IEnableDisablePosition {
     isEnabled: boolean;
@@ -42,20 +44,22 @@ export function loadRules(ruleConfiguration: {[name: string]: any},
     for (const ruleName in ruleConfiguration) {
         if (ruleConfiguration.hasOwnProperty(ruleName)) {
             const ruleValue = ruleConfiguration[ruleName];
-            const Rule = findRule(ruleName, rulesDirectories);
-            if (Rule == null) {
-                notFoundRules.push(ruleName);
-            } else {
-                if (isJs && Rule.metadata && Rule.metadata.typescriptOnly != null && Rule.metadata.typescriptOnly) {
-                    notAllowedInJsRules.push(ruleName);
+            if (AbstractRule.isRuleEnabled(ruleValue) || enableDisableRuleMap.hasOwnProperty(ruleName)) {
+                const Rule: (typeof AbstractRule) | null = findRule(ruleName, rulesDirectories);
+                if (Rule == null) {
+                    notFoundRules.push(ruleName);
                 } else {
-                    const ruleSpecificList = (ruleName in enableDisableRuleMap ? enableDisableRuleMap[ruleName] : []);
-                    const disabledIntervals = buildDisabledIntervalsFromSwitches(ruleSpecificList);
-                    rules.push(new Rule(ruleName, ruleValue, disabledIntervals));
+                    if (isJs && Rule.metadata && Rule.metadata.typescriptOnly != null && Rule.metadata.typescriptOnly) {
+                        notAllowedInJsRules.push(ruleName);
+                    } else {
+                        const ruleSpecificList = (ruleName in enableDisableRuleMap ? enableDisableRuleMap[ruleName] : []);
+                        const disabledIntervals = buildDisabledIntervalsFromSwitches(ruleSpecificList);
+                        rules.push(new (Rule as any)(ruleName, ruleValue, disabledIntervals));
 
-                    if (Rule.metadata && Rule.metadata.deprecationMessage && shownDeprecations.indexOf(Rule.metadata.ruleName) === -1) {
-                        console.warn(`${Rule.metadata.ruleName} is deprecated. ${Rule.metadata.deprecationMessage}`);
-                        shownDeprecations.push(Rule.metadata.ruleName);
+                        if (Rule.metadata && Rule.metadata.deprecationMessage && shownDeprecations.indexOf(Rule.metadata.ruleName) === -1) {
+                            console.warn(`${Rule.metadata.ruleName} is deprecated. ${Rule.metadata.deprecationMessage}`);
+                            shownDeprecations.push(Rule.metadata.ruleName);
+                        }
                     }
                 }
             }
@@ -89,26 +93,22 @@ export function loadRules(ruleConfiguration: {[name: string]: any},
 
 export function findRule(name: string, rulesDirectories?: string | string[]) {
     const camelizedName = transformName(name);
+    let Rule: typeof AbstractRule | null;
 
     // first check for core rules
-    let Rule = loadRule(CORE_RULES_DIRECTORY, camelizedName);
-    if (Rule != null) {
-        return Rule;
-    }
+    Rule = loadCachedRule(CORE_RULES_DIRECTORY, camelizedName);
 
-    const directories = getRulesDirectories(rulesDirectories);
-
-    for (const rulesDirectory of directories) {
+    if (Rule == null) {
         // then check for rules within the first level of rulesDirectory
-        if (rulesDirectory != null) {
-            Rule = loadRule(rulesDirectory, camelizedName);
+        for (const dir of arrayify(rulesDirectories)) {
+            Rule = loadCachedRule(dir, camelizedName, true);
             if (Rule != null) {
-                return Rule;
+                break;
             }
         }
     }
 
-    return undefined;
+    return Rule;
 }
 
 function transformName(name: string) {
@@ -127,15 +127,40 @@ function transformName(name: string) {
  */
 function loadRule(directory: string, ruleName: string) {
     const fullPath = path.join(directory, ruleName);
-
     if (fs.existsSync(fullPath + ".js")) {
         const ruleModule = require(fullPath);
         if (ruleModule && ruleModule.Rule) {
             return ruleModule.Rule;
         }
     }
-
     return undefined;
+}
+
+function loadCachedRule(directory: string, ruleName: string, isCustomPath = false) {
+    // use cached value if available
+    const fullPath = path.join(directory, ruleName);
+    const cachedRule = cachedRules.get(fullPath);
+    if (cachedRule !== undefined) {
+        return cachedRule;
+    }
+
+    // get absolute path
+    let absolutePath: string | undefined = directory;
+    if (isCustomPath) {
+        absolutePath = getRelativePath(directory);
+        if (absolutePath != null) {
+            if (!fs.existsSync(absolutePath)) {
+                throw new Error(`Could not find custom rule directory: ${directory}`);
+            }
+        }
+    }
+
+    let Rule: typeof AbstractRule | null = null;
+    if (absolutePath != null) {
+        Rule = loadRule(absolutePath, ruleName);
+    }
+    cachedRules.set(fullPath, Rule);
+    return Rule;
 }
 
 /**