+++ src/rules/memberOrderingRule.ts
@@ -19,6 +19,7 @@ import * as ts from "typescript";
 import * as Lint from "../index";
 
 const OPTION_ORDER = "order";
+const OPTION_ALPHABETIZE = "alphabetize";
 
 enum MemberKind {
     publicStaticField,
@@ -39,83 +40,6 @@ enum MemberKind {
 }
 
 const PRESETS = new Map<string, MemberCategoryJson[]>([
-    ["variables-before-functions", [
-        {
-            kinds: [
-                "public-static-field",
-                "protected-static-field",
-                "private-static-field",
-                "public-instance-field",
-                "protected-instance-field",
-                "private-instance-field",
-            ],
-            name: "field",
-        },
-        {
-            kinds: [
-                "constructor",
-                "public-static-method",
-                "protected-static-method",
-                "private-static-method",
-                "public-instance-method",
-                "protected-instance-method",
-                "private-instance-method",
-            ],
-            name: "method",
-        },
-    ]],
-    ["static-before-instance", [
-        {
-            kinds: [
-                "public-static-field",
-                "public-static-method",
-                "protected-static-field",
-                "protected-static-method",
-                "private-static-field",
-                "private-static-method",
-            ],
-            name: "static member",
-        },
-        {
-            kinds: [
-                "public-instance-field",
-                "protected-instance-field",
-                "private-instance-field",
-                "constructor",
-                "public-instance-method",
-                "protected-instance-method",
-                "private-instance-method",
-            ],
-            name: "instance member",
-        },
-    ]],
-    ["public-before-private", [
-        {
-            kinds: [
-                "public-static-field",
-                "protected-static-field",
-                "public-instance-field",
-                "protected-instance-field",
-                "public-instance-method",
-                "protected-instance-method",
-                "public-static-method",
-                "protected-static-method",
-                "public-constructor",
-                "protected-constructor",
-            ],
-            name: "public member",
-        },
-        {
-            kinds: [
-                "private-static-field",
-                "private-instance-field",
-                "private-instance-method",
-                "private-static-method",
-                "private-constructor",
-            ],
-            name: "private member",
-        },
-    ]],
     ["fields-first", [
         "public-static-field",
         "protected-static-field",
@@ -195,7 +119,9 @@ const optionsDescription = Lint.Utils.dedent`
                 "public-static-method",
                 "protected-static-method"
             ]
-        }`;
+        }
+
+    The '${OPTION_ALPHABETIZE}' option will enforce that members within the same category should be alphabetically sorted by name.`;
 
 export class Rule extends Lint.Rules.AbstractRule {
     /* tslint:disable:object-literal-sort-keys */
@@ -255,6 +181,14 @@ export class Rule extends Lint.Rules.AbstractRule {
         type: "typescript",
         typescriptOnly: true,
     };
+
+    public static FAILURE_STRING_ALPHABETIZE(prevName: string, curName: string) {
+        return `${show(curName)} should come alphabetically before ${show(prevName)}`;
+        function show(s: string) {
+            return s === "" ? "Computed property" : `'${s}'`;
+        }
+    }
+
     /* tslint:enable:object-literal-sort-keys */
     public apply(sourceFile: ts.SourceFile): Lint.RuleFailure[] {
         return this.applyWithWalker(new MemberOrderingWalker(sourceFile, this.getOptions()));
@@ -262,11 +196,11 @@ export class Rule extends Lint.Rules.AbstractRule {
 }
 
 export class MemberOrderingWalker extends Lint.RuleWalker {
-    private readonly order: MemberCategory[];
+    private readonly opts: Options;
 
     constructor(sourceFile: ts.SourceFile, options: Lint.IOptions) {
         super(sourceFile, options);
-        this.order = getOrder(this.getOptions());
+        this.opts = parseOptions(this.getOptions());
     }
 
     public visitClassDeclaration(node: ts.ClassDeclaration) {
@@ -291,6 +225,7 @@ export class MemberOrderingWalker extends Lint.RuleWalker {
 
     private visitMembers(members: Member[]) {
         let prevRank = -1;
+        let prevName: string | undefined;
         for (const member of members) {
             const rank = this.memberRank(member);
             if (rank === -1) {
@@ -309,10 +244,39 @@ export class MemberOrderingWalker extends Lint.RuleWalker {
                     `Instead, this should come ${locationHint}.`;
                 this.addFailureAtNode(member, errorLine1);
             } else {
+                if (this.opts.alphabetize && member.name) {
+                    if (rank !== prevRank) {
+                        // No alphabetical ordering between different ranks
+                        prevName = undefined;
+                    }
+
+                    const curName = nameString(member.name);
+                    if (prevName !== undefined && caseInsensitiveLess(curName, prevName)) {
+                        this.addFailureAtNode(member.name,
+                            Rule.FAILURE_STRING_ALPHABETIZE(this.findLowerName(members, rank, curName), curName));
+                    } else {
+                        prevName = curName;
+                    }
+                }
+
                 // keep track of last good node
                 prevRank = rank;
             }
         }
+    }
+
+    /** Finds the lowest name higher than 'targetName'. */
+    private findLowerName(members: Member[], targetRank: Rank, targetName: string): string {
+        for (const member of members) {
+            if (!member.name || this.memberRank(member) !== targetRank) {
+                continue;
+            }
+            const name = nameString(member.name);
+            if (caseInsensitiveLess(targetName, name)) {
+                return name;
+            }
+        }
+        throw new Error("Expected to find a name");
     }
 
     /** Finds the highest existing rank lower than `targetRank`. */
@@ -332,12 +296,16 @@ export class MemberOrderingWalker extends Lint.RuleWalker {
         if (optionName === undefined) {
             return -1;
         }
-        return this.order.findIndex((category) => category.has(optionName));
+        return this.opts.order.findIndex((category) => category.has(optionName));
     }
 
     private rankName(rank: Rank): string {
-        return this.order[rank].name;
+        return this.opts.order[rank].name;
     }
+}
+
+function caseInsensitiveLess(a: string, b: string) {
+    return a.toLowerCase() < b.toLowerCase();
 }
 
 function memberKindForConstructor(access: Access): MemberKind {
@@ -406,22 +374,30 @@ type Rank = number;
 
 type Access = "public" | "protected" | "private";
 
-function getOrder(options: any[]): MemberCategory[] {
-    return getOrderJson(options).map((cat) => typeof cat === "string"
+interface Options {
+    order: MemberCategory[];
+    alphabetize: boolean;
+}
+
+function parseOptions(options: any[]): Options {
+    const { order: orderJson, alphabetize } = getOptionsJson(options);
+    const order = orderJson.map((cat) => typeof cat === "string"
         ? new MemberCategory(cat.replace(/-/g, " "), new Set(memberKindFromName(cat)))
         : new MemberCategory(cat.name, new Set(flatMap(cat.kinds, memberKindFromName))));
+    return { order, alphabetize };
 }
-function getOrderJson(allOptions: any[]): MemberCategoryJson[] {
+function getOptionsJson(allOptions: any[]): { order: MemberCategoryJson[], alphabetize: boolean } {
     if (allOptions == null || allOptions.length === 0 || allOptions[0] == null) {
         throw new Error("Got empty options");
     }
 
     const firstOption = allOptions[0];
     if (typeof firstOption !== "object") {
-        return categoryFromOption(firstOption);
+        // Undocumented direct string option. Deprecate eventually.
+        return { order: convertFromOldStyleOptions(allOptions), alphabetize: false }; // presume allOptions to be string[]
     }
 
-    return categoryFromOption(firstOption[OPTION_ORDER]);
+    return { order: categoryFromOption(firstOption[OPTION_ORDER]), alphabetize: !!firstOption[OPTION_ALPHABETIZE] };
 }
 function categoryFromOption(orderOption: {}): MemberCategoryJson[] {
     if (Array.isArray(orderOption)) {
@@ -435,6 +411,53 @@ function categoryFromOption(orderOption: {}): MemberCategoryJson[] {
     return preset;
 }
 
+/**
+ * Convert from undocumented old-style options.
+ * This is designed to mimic the old behavior and should be removed eventually.
+ */
+function convertFromOldStyleOptions(options: string[]): MemberCategoryJson[] {
+    let categories: NameAndKinds[] = [{ name: "member", kinds: allMemberKindNames }];
+    if (hasOption("variables-before-functions")) {
+        categories = splitOldStyleOptions(categories, (kind) => kind.includes("field"), "field", "method");
+    }
+    if (hasOption("static-before-instance")) {
+        categories = splitOldStyleOptions(categories, (kind) => kind.includes("static"), "static", "instance");
+    }
+    if (hasOption("public-before-private")) {
+        // 'protected' is considered public
+        categories = splitOldStyleOptions(categories, (kind) => !kind.includes("private"), "public", "private");
+    }
+    return categories;
+
+    function hasOption(x: string): boolean {
+        return options.indexOf(x) !== -1;
+    }
+}
+interface NameAndKinds { name: string; kinds: string[]; }
+function splitOldStyleOptions(categories: NameAndKinds[], filter: (name: string) => boolean, a: string, b: string): NameAndKinds[] {
+    const newCategories: NameAndKinds[]  = [];
+    for (const cat of categories) {
+        const yes = []; const no = [];
+        for (const kind of cat.kinds) {
+            if (filter(kind)) {
+                yes.push(kind);
+            } else {
+                no.push(kind);
+            }
+        }
+        const augmentName = (s: string) => {
+            if (a === "field") {
+                // Replace "member" with "field"/"method" instead of augmenting.
+                return s;
+            }
+            return `${s} ${cat.name}`;
+        };
+        newCategories.push({ name: augmentName(a), kinds: yes });
+        newCategories.push({ name: augmentName(b), kinds: no });
+    }
+    return newCategories;
+}
+
 function isFunctionLiteral(node: ts.Node | undefined) {
     switch (node && node.kind) {
         case ts.SyntaxKind.ArrowFunction:
@@ -442,6 +465,17 @@ function isFunctionLiteral(node: ts.Node | undefined) {
             return true;
         default:
             return false;
+    }
+}
+
+function nameString(name: ts.PropertyName): string {
+    switch (name.kind) {
+        case ts.SyntaxKind.Identifier:
+        case ts.SyntaxKind.StringLiteral:
+        case ts.SyntaxKind.NumericLiteral:
+            return (name as ts.Identifier | ts.LiteralExpression).text;
+        default:
+            return "";
     }
 }